+++ core/src/main/java/nl/nn/adapterframework/configuration/WebContentPropertyPlaceholderConfigurer.java
@@ -1,0 +1,54 @@
+/*
+   Copyright 2016 Nationale-Nederlanden
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+ */
+package nl.nn.adapterframework.configuration;
+
+import java.util.Properties;
+
+import nl.nn.adapterframework.util.AppConstants;
+import nl.nn.adapterframework.util.Misc;
+
+import org.apache.commons.lang.StringUtils;
+import org.springframework.beans.factory.config.PropertyPlaceholderConfigurer;
+
+/**
+ * Make the webContent.dir property available to the Ibis configuration and the
+ * Spring configuration.
+ * 
+ * @author Peter Leeuwenburgh
+ */
+public class WebContentPropertyPlaceholderConfigurer extends
+		PropertyPlaceholderConfigurer {
+	private static String WEBCONTENT_DIR_PROPERTY = "webContent.dir";
+
+	public WebContentPropertyPlaceholderConfigurer() {
+		setIgnoreUnresolvablePlaceholders(true);
+	}
+
+	@Override
+	protected void convertProperties(Properties props) {
+		AppConstants appConstants = AppConstants.getInstance();
+		String webContentDir = appConstants
+				.getResolvedProperty(WEBCONTENT_DIR_PROPERTY);
+		if (StringUtils.isEmpty(webContentDir)) {
+			webContentDir = Misc.getWebContentDirectory();
+			if (webContentDir != null) {
+				appConstants.setPropertyPlaceholderConfigurerProperty(
+						WEBCONTENT_DIR_PROPERTY, webContentDir);
+				props.put(WEBCONTENT_DIR_PROPERTY, webContentDir);
+			}
+		}
+	}
+}